requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
gspread>=5.0.0
google-auth>=2.0.0
plotly>=5.0.0
openpyxl>=3.0.0
pyarrow>=12.0.0  # For faster file processing